Allow non integer qty_increments in backend config A wrong value handling was in place not allowing qty_increments to be a float - Fixed wrong JS validation forcing the qty_increment to be an integer - Fixed Stock Item model to work with flaot values too

+++ app/code/Magento/CatalogInventory/Model/Stock/Item.php
@@ -401,7 +401,7 @@ class Item extends AbstractExtensibleModel implements StockItemInterface
                 if ($this->getUseConfigQtyIncrements()) {
                     $this->qtyIncrements = $this->stockConfiguration->getQtyIncrements($this->getStoreId());
                 } else {
-                    $this->qtyIncrements = (int) $this->getData(static::QTY_INCREMENTS);
+                    $this->qtyIncrements = (float) $this->getData(static::QTY_INCREMENTS);
                 }
             }
             if ($this->qtyIncrements <= 0) {
